makefile

+++ ft_bzero.c
@@ -6,7 +6,7 @@
 /*   By: hboutale <hboutale@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/10/30 13:05:56 by hboutale          #+#    #+#             */
-/*   Updated: 2024/10/31 14:06:27 by hboutale         ###   ########.fr       */
+/*   Updated: 2024/11/03 11:31:05 by hboutale         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 void	ft_bzero(void *ptr, size_t size)
 {
-	int i;
+	size_t i;
 	unsigned char *str;
 
 	str = ptr;
