+++ ft_memchr.c
@@ -6,7 +6,7 @@
 /*   By: hboutale <hboutale@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/10/30 17:33:33 by hboutale          #+#    #+#             */
-/*   Updated: 2024/10/31 14:45:27 by hboutale         ###   ########.fr       */
+/*   Updated: 2024/11/03 11:32:13 by hboutale         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -16,7 +16,7 @@ void	*ft_memchr(void *ptr, int c, size_t n)
 {
 	unsigned char *str;
 	unsigned char ch;
-	int i;
+	size_t i;
 
 	str = ptr;
 	ch = c;
